+++ Elements/src/BIM/Door/Door.cs
@@ -11,37 +11,43 @@ namespace Elements
         /// <summary>Definition of a door</summary>
         public class Door : GeometricElement
         {
+                /// <summary>The material to be used on the door frame</summary>
                 public Material FrameMaterial { get; set; } = new Material(Colors.Gray, 0.5, 0.25, false, null, false, false, null, false, null, 0, false, default, "Silver Frame");
-
                 /// <summary>The opening type of the door that should be placed</summary>
                 [JsonProperty("Door Opening Type")]
+                [Newtonsoft.Json.JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
                 public DoorOpeningType OpeningType { get; private set; }
                 /// <summary>The opening side of the door that should be placed</summary>
                 [JsonProperty("Door Opening Side")]
+                [Newtonsoft.Json.JsonConverter(typeof(Newtonsoft.Json.Converters.StringEnumConverter))]
                 public DoorOpeningSide OpeningSide { get; private set; }
                 /// <summary>Width of a door without a frame.</summary>
+                [JsonProperty("Door Width")]
                 public double DoorWidth { get; set; }
                 /// <summary>Height of a door without a frame.</summary>
+                [JsonProperty("Door Height")]
                 public double DoorHeight { get; set; }
+                /// <summary>Type of door (Glass or Solid).</summary>
+                [JsonProperty("Door Type")]
+                public string DoorType { get; set; }
                 /// <summary>Default door thickness.</summary>
-                public static double DEFAULT_DOOR_THICKNESS = 2 * 0.0254;
+                public static double DEFAULT_DOOR_THICKNESS = Units.InchesToMeters(2.0);
                 /// <summary>Door thickness.</summary>
                 public double DoorThickness { get; set; } = DEFAULT_DOOR_THICKNESS;
                 /// <summary>Default thickness of a door frame.</summary>
-                public double FrameDepth { get; set; } = 4 * 0.0254;
+                public double FrameDepth { get; set; } = Units.InchesToMeters(4.0);
                 /// <summary>Default width of a door frame.</summary>
-                public double FrameWidth { get; set; } = 2 * 0.0254; //2 inches
-
+                public double FrameWidth { get; set; } = Units.InchesToMeters(2.0); //2 inches
                 /// <summary>Height of the door handle from the ground</summary>
-                public double HandleHeight { get; set; } = 42 * 0.0254;
+                public double HandleHeight { get; set; } = Units.InchesToMeters(42.0);
                 /// <summary>Radius of the fixture against the door</summary>
-                public double HandleBaseRadius { get; set; } = 1.35 * 0.0254;
+                public double HandleBaseRadius { get; set; } = Units.InchesToMeters(1.35);
                 /// <summary>Radius of the handle</summary>
-                public double HandleRadius { get; set; } = 0.45 * 0.0254;
+                public double HandleRadius { get; set; } = Units.InchesToMeters(0.45);
                 /// <summary>Length of the handle</summary>
-                public double HandleLength { get; set; } = 5 * 0.0254;
+                public double HandleLength { get; set; } = Units.InchesToMeters(5.0);
                 /// <summary>Depth of the handle from the face of the door</summary>
-                public double HandleDepth { get; set; } = 2 * 0.0254;
+                public double HandleDepth { get; set; } = Units.InchesToMeters(2.0);
                 /// <summary>Original position of the door used for override identity</summary>
                 public Vector3 OriginalPosition { get; set; }
 
@@ -186,9 +192,10 @@ namespace Elements
                         {
                                 this.CreateDoorSolidRepresentation(),
                                 this.CreateDoorFrameRepresentation(),
-                                this.CreateDoorCurveRepresentation(),
                                 this.CreateDoorHandleRepresentation()
                         };
+
+                        representationInstances.AddRange(this.CreateDoorCurveRepresentation());
 
                         return representationInstances.Where(instance => instance != null).ToList();
                 }
@@ -216,14 +223,11 @@ namespace Elements
                         return 0;
                 }
 
-                private RepresentationInstance CreateDoorCurveRepresentation()
+                private List<RepresentationInstance> CreateDoorCurveRepresentation()
                 {
-                        var points = CollectPointsForSchematicVisualization();
-                        var curve = new IndexedPolycurve(points);
-                        var curveRep = new CurveRepresentation(curve, false);
-                        var repInstance = new RepresentationInstance(curveRep, BuiltInMaterials.Black);
+                        var repInstances = CollectPointsForSchematicVisualization();
 
-                        return repInstance;
+                        return repInstances;
                 }
 
                 private RepresentationInstance CreateDoorFrameRepresentation()
@@ -251,6 +255,7 @@ namespace Elements
                         var doorFrameExtrude = new Extrude(new Profile(doorFramePolygon), this.FrameDepth, Vector3.YAxis);
 
                         var solidRep = new SolidRepresentation(doorFrameExtrude);
+                        solidRep.SetSnappingPoints(new List<SnappingPoints>());
                         var repInstance = new RepresentationInstance(solidRep, FrameMaterial, true);
                         return repInstance;
                 }
@@ -286,45 +291,69 @@ namespace Elements
                         }
 
                         var solidRep = new SolidRepresentation(doorExtrusions);
+                        solidRep.SetSnappingPoints(new List<SnappingPoints>(new SnappingPoints[] { new SnappingPoints(new[] { left, Vector3.Origin, right }, SnappingEdgeMode.LineStrip) }));
                         var repInstance = new RepresentationInstance(solidRep, this.Material, true);
                         return repInstance;
                 }
 
-                private List<Vector3> CollectPointsForSchematicVisualization()
+                private List<RepresentationInstance> CollectPointsForSchematicVisualization()
                 {
-                        var points = new List<Vector3>();
+                        var representationInstances = new List<RepresentationInstance>();
 
                         if (this.OpeningSide == DoorOpeningSide.Undefined || this.OpeningType == DoorOpeningType.Undefined)
                         {
-                                return points;
+                                return representationInstances;
                         }
 
                         if (this.OpeningSide != DoorOpeningSide.LeftHand)
                         {
-                                points.AddRange(CollectSchematicVisualizationLines(this, false, false, 90));
+                                var points = CollectSchematicVisualizationLines(this, false, false, 90);
+                                points.Add(points[0]);
+                                var curve = new IndexedPolycurve(points);
+                                var curveRep = new CurveRepresentation(curve, false);
+                                curveRep.SetSnappingPoints(new List<SnappingPoints>());
+                                var repInstance = new RepresentationInstance(curveRep, BuiltInMaterials.Black);
+                                representationInstances.Add(repInstance);
                         }
 
                         if (this.OpeningSide != DoorOpeningSide.RightHand)
                         {
-                                points.AddRange(CollectSchematicVisualizationLines(this, true, false, 90));
+                                var points = CollectSchematicVisualizationLines(this, true, false, 90);
+                                points.Add(points[0]);
+                                var curve = new IndexedPolycurve(points);
+                                var curveRep = new CurveRepresentation(curve, false);
+                                curveRep.SetSnappingPoints(new List<SnappingPoints>());
+                                var repInstance = new RepresentationInstance(curveRep, BuiltInMaterials.Black);
+                                representationInstances.Add(repInstance);
                         }
 
-                        if (this.OpeningType == DoorOpeningType.SingleSwing)
+                        if (this.OpeningType == DoorOpeningType.DoubleSwing)
                         {
-                                return points;
+
+                                if (this.OpeningSide != DoorOpeningSide.LeftHand)
+                                {
+                                        var points = CollectSchematicVisualizationLines(this, false, true, 90);
+                                        points.Add(points[0]);
+                                        var curve = new IndexedPolycurve(points);
+                                        var curveRep = new CurveRepresentation(curve, false);
+                                        curveRep.SetSnappingPoints(new List<SnappingPoints>());
+                                        var repInstance = new RepresentationInstance(curveRep, BuiltInMaterials.Black);
+                                        representationInstances.Add(repInstance);
+                                }
+
+                                if (this.OpeningSide != DoorOpeningSide.RightHand)
+                                {
+                                        var points = CollectSchematicVisualizationLines(this, true, true, 90);
+                                        points.Add(points[0]);
+                                        var curve = new IndexedPolycurve(points);
+                                        var curveRep = new CurveRepresentation(curve, false);
+                                        curveRep.SetSnappingPoints(new List<SnappingPoints>());
+                                        var repInstance = new RepresentationInstance(curveRep, BuiltInMaterials.Black);
+                                        representationInstances.Add(repInstance);
+                                }
                         }
 
-                        if (this.OpeningSide != DoorOpeningSide.LeftHand)
-                        {
-                                points.AddRange(CollectSchematicVisualizationLines(this, false, true, 90));
-                        }
-
-                        if (this.OpeningSide != DoorOpeningSide.RightHand)
-                        {
-                                points.AddRange(CollectSchematicVisualizationLines(this, true, true, 90));
-                        }
-
-                        return points;
+                        return representationInstances;
                 }
 
                 private List<Vector3> CollectSchematicVisualizationLines(Door door, bool leftSide, bool inside, double angle)
@@ -402,32 +431,34 @@ namespace Elements
 
                         if (OpeningSide == DoorOpeningSide.DoubleDoor)
                         {
-                                var handlePair1 = CreateHandlePair(-3 * 0.0254, false);
+                                var handlePair1 = CreateHandlePair(DoorWidth / 2 + Units.InchesToMeters(3.0), false);
                                 solidOperationsList.AddRange(handlePair1);
 
-                                var handlePair2 = CreateHandlePair(3 * 0.0254, true);
+                                var handlePair2 = CreateHandlePair(DoorWidth / 2 + Units.InchesToMeters(3.0), true);
                                 solidOperationsList.AddRange(handlePair2);
                         }
                         else if (OpeningSide != DoorOpeningSide.Undefined)
                         {
-                                var xPos = OpeningSide == DoorOpeningSide.LeftHand ? -(FullDoorWidthWithoutFrame / 2 - 2 * 0.0254) : (FullDoorWidthWithoutFrame / 2 - 2 * 0.0254);
+                                var xPos = OpeningSide == DoorOpeningSide.LeftHand ? Units.InchesToMeters(3.0) : Units.InchesToMeters(3.0);
                                 var handle = CreateHandlePair(xPos, OpeningSide == DoorOpeningSide.LeftHand);
                                 solidOperationsList.AddRange(handle);
                         }
 
                         var solidRep = new SolidRepresentation(solidOperationsList);
+                        solidRep.SetSnappingPoints(new List<SnappingPoints>());
                         var repInst = new RepresentationInstance(solidRep, FrameMaterial);
                         return repInst;
                 }
 
-                private List<SolidOperation> CreateHandlePair(double xRelPos, bool isCodirectionalToX)
+                private List<SolidOperation> CreateHandlePair(double handleOffset, bool isCodirectionalToX)
                 {
-                        var xOffset = xRelPos * DoorWidth * Vector3.XAxis;
-                        var yOffset = DoorThickness * Vector3.YAxis;
+                        var handleDir = isCodirectionalToX ? Vector3.XAxis : Vector3.XAxis.Negate();
+
+                        var xOffset = (-DoorWidth / 2 + handleOffset) * handleDir;
+                        var yOffset = DoorThickness / 2 * Vector3.YAxis;
                         var zOffset = HandleHeight * Vector3.ZAxis;
 
                         var solidOperationsList = new List<SolidOperation>();
-                        var handleDir = isCodirectionalToX ? Vector3.XAxis : Vector3.XAxis.Negate();
 
                         var handleOrigin1 = xOffset + yOffset + zOffset;
                         var handle1Ops = CreateHandle(handleOrigin1, handleDir, Vector3.YAxis);
@@ -444,9 +475,9 @@ namespace Elements
                 {
                         var circleTransform = new Transform(origin, handleDir, yDir);
                         var circle = new Circle(circleTransform, HandleBaseRadius).ToPolygon();
-                        var circleOperation = new Extrude(circle, 0.1 * HandleDepth, yDir);
+                        var circleOperation = new Extrude(circle, 0.1 * HandleDepth, yDir.Negate());
 
-                        var cyl1Transform = new Transform(origin + 0.1 * HandleDepth * yDir, handleDir, yDir);
+                        var cyl1Transform = new Transform(origin, handleDir, yDir);
                         var cyl1Circle = new Circle(cyl1Transform, HandleRadius).ToPolygon();
                         var cyl1Operation = new Extrude(cyl1Circle, 0.9 * HandleDepth, yDir);